#32 test missing documentation and package reorganization
