+++ src/test/java/com/elbraulio/ezload/exception/EzExceptionTest.java
@@ -5,6 +5,8 @@ import org.hamcrest.MatcherAssert;
 import org.junit.Test;
 
 /**
+ * Unit test for {@link EzException}.
+ *
  * @author Braulio Lopez (brauliop.3@gmail.com)
  */
 public class EzExceptionTest {
