+++ src/test/java/com/elbraulio/ezload/sql/SqlFromParserTest.java
@@ -10,6 +10,8 @@ import org.hamcrest.MatcherAssert;
 import org.junit.Test;
 
 /**
+ * Unit test for {@link SqlFromParser}.
+ *
  * @author Braulio Lopez (brauliop.3@gmail.com)
  */
 public class SqlFromParserTest {
