+++ src/test/java/com/elbraulio/ezload/transform/ToStringTest.java
@@ -1,4 +1,4 @@
-package com.elbraulio.ezload.model;
+package com.elbraulio.ezload.transform;
 
 import com.elbraulio.ezload.transform.ToString;
 import org.hamcrest.CoreMatchers;
@@ -6,6 +6,8 @@ import org.hamcrest.MatcherAssert;
 import org.junit.Test;
 
 /**
+ * Unit test for {@link ToString}.
+ *
  * @author Braulio Lopez (brauliop.3@gmail.com)
  */
 public class ToStringTest {
